smsapp: Make messages in ConversationDisplay always stick to the bottom
This commit mostly reverses the direction of the ListView used by ConversationDisplay to make it bottom-to-top. This matches the way most chat apps work today (e.g. Discord), by leaving an empty space at the top when there are not enough messages to fill the screen.

Testing it, it seems like a bunch of the custom logic for scrolling down when there are new messages and staying focused on the current message while loading new ones can now be transparently handled by the ListView, so this trims down the code quite a lot.

While re-working the message loading, I also switched it to use a constant threshold rather than a proportional one, which fixes cases where scrolling down could load older messages.

[Enregistrement d'écran_20250707_103723.webm](/uploads/99efd26958e73fed1ce6f0e04bae51f4/Enregistrement_d_%C3%A9cran_20250707_103723.webm)

Overall, this makes the ConversationDisplay scrolling much smoother.